feat: refactir for multiple networks, add hypeEVM
