+++ src/app/config/constants.ts
@@ -1,0 +1,26 @@
+export const POINTS_ID_ETHENA_SATS_S3 = "POINTS_ID_ETHENA_SATS_S3";
+export const POINTS_ID_ETHENA_SATS_S4 = "POINTS_ID_ETHENA_SATS_S4";
+export const POINTS_ID_KARAK_S2 = "POINTS_ID_KARAK_S2";
+export const POINTS_ID_SYMBIOTIC_S1 = "POINTS_ID_SYMBIOTIC_S1";
+export const POINTS_ID_EIGENLAYER_S3 = "POINTS_ID_EIGENLAYER_S3";
+export const POINTS_ID_EIGENPIE_S1 = "POINTS_ID_EIGENPIE_S1";
+export const POINTS_ID_MELLOW_S1 = "POINTS_ID_MELLOW_S1";
+export const POINTS_ID_MERITS_S1 = "POINTS_ID_MERITS_S1";
+export const POINTS_ID_ZIRCUIT_S3 = "POINTS_ID_ZIRCUIT_S3";
+export const POINTS_ID_ETHERFI_S4 = "POINTS_ID_ETHERFI_S4";
+export const POINTS_ID_ETHERFI_S5 = "POINTS_ID_ETHERFI_S5";
+export const POINTS_ID_VEDA_S1 = "POINTS_ID_VEDA_S1";
+export const POINTS_ID_LOMBARD_LUX_S1 = "POINTS_ID_LOMBARD_LUX_S1";
+export const POINTS_ID_RESOLV_S1 = "POINTS_ID_RESOLV_S1";
+
+// HypeEVM
+export const POINTS_ID_HYPERBEAT_S1 = "POINTS_ID_HYPERBEAT_S1";
+export const POINTS_ID_HYPERLEND_S1 = "POINTS_ID_HYPERLEND_S1";
+export const POINTS_ID_HYPURRFI_S1 = "POINTS_ID_HYPURRFI_S1";
+export const POINTS_ID_HYPERSWAP_S1 = "POINTS_ID_HYPERSWAP_S1";
+export const POINTS_ID_TIMESWAP_S1 = "POINTS_ID_TIMESWAP_S1";
+export const POINTS_ID_SILHOUETTE_S1 = "POINTS_ID_SILHOUETTE_S1";
+export const POINTS_ID_UPSHIFT_S1 = "POINTS_ID_UPSHIFT_S1";
+export const POINTS_ID_SENTIMENT_S1 = "POINTS_ID_SENTIMENT_S1";
+
+export const MAINNET_AGETH = "0xe1B4d34E8754600962Cd944B535180Bd758E6c2e";
