+++ src/app/page.tsx
@@ -4,7 +4,9 @@ import Image from "next/image";
 import React, { useState, useEffect, useRef } from "react";
 import * as Tooltip from "@radix-ui/react-tooltip";
 import * as Dialog from "@radix-ui/react-dialog";
-import { CONFIG } from "./lib";
+import * as ToggleGroup from "@radix-ui/react-toggle-group";
+import { ETH_CONFIG } from "@/app/config/ethStrategies";
+import { HYPE_EVM_CONFIG } from "@/app/config/hypeEvmStrategies";
 
 type HistoricalData = {
   actualPoints: string;
@@ -292,6 +294,8 @@ export default function PointsAuditByPointsId() {
   const [data, setData] = useState<any[]>([]);
   const [loading, setLoading] = useState<boolean>(true);
   const [error, setError] = useState<string | null>(null);
+  const [selectedNetwork, setSelectedNetwork] = useState<string>("ethereum");
+  const [config, setConfig] = useState(ETH_CONFIG);
   const [historicalData, setHistoricalData] = useState<
     Record<string, HistoricalData>
   >({});
@@ -305,10 +309,16 @@ export default function PointsAuditByPointsId() {
     pointsId: "",
   });
 
+  // Update config when network changes
+  useEffect(() => {
+    setConfig(selectedNetwork === "ethereum" ? ETH_CONFIG : HYPE_EVM_CONFIG);
+    setLoading(true);
+  }, [selectedNetwork]);
+
   // Fetch logic
   useEffect(() => {
     Promise.all([
-      fetch("/api/points-audit").then((r) => r.json()),
+      fetch(`/api/points-audit/${selectedNetwork}`).then((r) => r.json()),
       fetch("/api/points-audit-history").then((r) => r.json()),
     ])
       .then(([pointsData, historyData]) => {
@@ -320,7 +330,7 @@ export default function PointsAuditByPointsId() {
         setError(err.message);
         setLoading(false);
       });
-  }, []);
+  }, [selectedNetwork]);
 
   // Error display
   if (error) {
@@ -336,457 +346,476 @@ export default function PointsAuditByPointsId() {
   // Grouping the data once we have it
   const groupedData = groupByPointsId(data);
 
-  /**
-   * -------------------------
-   * 1) LOADING State
-   * -------------------------
-   * We'll handle a skeleton for both table and the card layout.
-   */
-  if (loading) {
-    return (
-      <div className="p-8 max-w-7xl mx-auto" aria-busy="true" role="status">
-        {/* TABLE SKELETON (Visible on md screens and up) */}
-        <div className="hidden sm:block overflow-x-auto rounded-lg shadow-lg border border-gray-200">
-          <table className="w-full border-collapse bg-white">
-            <thead>
-              <tr className="bg-gray-50">
-                <th
-                  scope="col"
-                  className="p-4 text-left text-gray-700 font-semibold border-b border-gray-200"
-                >
-                  Points Type
-                </th>
-                <th
-                  scope="col"
-                  className="p-4 text-left text-gray-700 font-semibold border-b border-gray-200"
-                >
-                  Strategy
-                </th>
-                <th
-                  scope="col"
-                  className="p-4 text-right text-gray-700 font-semibold border-b border-gray-200"
-                >
-                  Expected
-                </th>
-                <th
-                  scope="col"
-                  className="p-4 text-right text-gray-700 font-semibold border-b border-gray-200"
-                >
-                  Actual
-                </th>
-                <th
-                  scope="col"
-                  className="p-4 text-right text-gray-700 font-semibold border-b border-gray-200"
-                >
-                  Diff
-                </th>
-                <th
-                  scope="col"
-                  className="p-4 text-right text-gray-700 font-semibold border-b border-gray-200"
-                >
-                  Audit State
-                </th>
-              </tr>
-            </thead>
-            <tbody>
-              {Array.from({ length: 5 }).map((_, groupIndex) => (
-                <React.Fragment key={groupIndex}>
-                  <tr className="border-b bg-gray-50">
-                    <td
-                      className="p-4 font-medium animate-pulse bg-gray-200 text-gray-200"
-                      colSpan={6}
-                    >
-                      <div className="h-4 w-48 bg-gray-200 rounded"></div>
-                    </td>
-                  </tr>
-                  {Array.from({ length: 3 }).map((__, rowIndex) => (
-                    <tr
-                      key={`${groupIndex}-${rowIndex}`}
-                      className="border-b hover:bg-gray-50 transition-colors"
-                    >
-                      <td className="p-4" />
-                      <td className="p-4">
-                        <div className="h-4 w-32 bg-gray-200 rounded animate-pulse mb-2" />
-                      </td>
-                      <td className="p-4 text-right">
-                        <div className="h-4 w-16 bg-gray-200 rounded animate-pulse ml-auto" />
-                      </td>
-                      <td className="p-4 text-right">
-                        <div className="h-4 w-16 bg-gray-200 rounded animate-pulse ml-auto" />
-                      </td>
-                      <td className="p-4 text-right">
-                        <div className="h-4 w-16 bg-gray-200 rounded animate-pulse ml-auto" />
-                      </td>
-                    </tr>
-                  ))}
-                </React.Fragment>
-              ))}
-            </tbody>
-          </table>
-        </div>
-
-        {/* CARD SKELETON (Visible on mobile screens) */}
-        <div className="block sm:hidden space-y-4">
-          {Array.from({ length: 5 }).map((_, idx) => (
-            <div
-              key={idx}
-              className="bg-white border border-gray-200 rounded-lg p-4 animate-pulse"
-            >
-              <div className="h-4 w-32 bg-gray-200 rounded mb-4" />
-              {Array.from({ length: 3 }).map((__, rIdx) => (
-                <div key={rIdx} className="flex justify-between mb-2">
-                  <div className="h-4 w-20 bg-gray-200 rounded" />
-                  <div className="h-4 w-16 bg-gray-200 rounded" />
-                </div>
-              ))}
-            </div>
-          ))}
-        </div>
-      </div>
-    );
-  }
-
-  /**
-   * -------------------------
-   * 2) ACTUAL DATA RENDER
-   * -------------------------
-   */
   return (
     <div className="p-8 max-w-7xl mx-auto">
-      {/* ---------- TABLE Layout (hidden on small screens) ---------- */}
-      <div className="hidden sm:block overflow-x-auto rounded-lg shadow-lg border border-gray-200">
-        <table
-          className="w-full border-collapse bg-white"
-          aria-label="Points Audit Table"
+      <div className="mb-8">
+        <ToggleGroup.Root
+          type="single"
+          value={selectedNetwork}
+          onValueChange={(value: string | undefined) =>
+            value && setSelectedNetwork(value)
+          }
+          className="inline-flex rounded-lg border border-gray-200 bg-gray-50 p-1 shadow-md"
         >
-          <thead>
-            <tr className="bg-gray-50">
-              <th
-                scope="col"
-                className="p-4 text-left text-gray-700 font-semibold border-b border-gray-200 w-[60px]"
-              ></th>
-              <th
-                scope="col"
-                className="p-4 text-left text-gray-700 font-semibold border-b border-gray-200"
-              >
-                Strategy
-              </th>
-              <th
-                scope="col"
-                className="p-4 text-right text-gray-700 font-semibold border-b border-gray-200 w-[140px]"
-              >
-                Expected
-              </th>
-              <th
-                scope="col"
-                className="p-4 text-right text-gray-700 font-semibold border-b border-gray-200 w-[140px]"
-              >
-                Actual
-              </th>
-              <th
-                scope="col"
-                className="p-4 text-right text-gray-700 font-semibold border-b border-gray-200 w-[160px]"
-              >
-                Diff
-              </th>
-              <th
-                scope="col"
-                className="p-4 text-right text-gray-700 font-semibold border-b border-gray-200 w-[160px]"
-              >
-                <div>Audit State</div>
-                <div className="text-xs font-normal">Last Snapshot, Diff%</div>
-              </th>
-            </tr>
-          </thead>
+          <ToggleGroup.Item
+            value="ethereum"
+            className="px-4 py-2 text-gray-700 font-medium rounded-md data-[state=on]:font-semibold data-[state=on]:bg-gray-300 data-[state=on]:text-gray-900 data-[state=on]:shadow-sm hover:bg-gray-100 transition-colors"
+          >
+            Ethereum
+          </ToggleGroup.Item>
+          <ToggleGroup.Item
+            value="hypeEVM"
+            className="px-4 py-2 text-gray-700 font-medium rounded-md data-[state=on]:font-semibold data-[state=on]:bg-gray-300 data-[state=on]:text-gray-900 data-[state=on]:shadow-sm hover:bg-gray-100 transition-colors"
+          >
+            HypeEVM
+          </ToggleGroup.Item>
+        </ToggleGroup.Root>
+      </div>
 
-          <tbody>
+      {loading ? (
+        <>
+          {/* TABLE SKELETON (Visible on md screens and up) */}
+          <div className="hidden sm:block overflow-x-auto rounded-lg shadow-lg border border-gray-200">
+            <table className="w-full border-collapse bg-white">
+              <thead>
+                <tr className="bg-gray-50">
+                  <th
+                    scope="col"
+                    className="p-4 text-left text-gray-700 font-semibold border-b border-gray-200"
+                  >
+                    Points Type
+                  </th>
+                  <th
+                    scope="col"
+                    className="p-4 text-left text-gray-700 font-semibold border-b border-gray-200"
+                  >
+                    Strategy
+                  </th>
+                  <th
+                    scope="col"
+                    className="p-4 text-right text-gray-700 font-semibold border-b border-gray-200"
+                  >
+                    Expected
+                  </th>
+                  <th
+                    scope="col"
+                    className="p-4 text-right text-gray-700 font-semibold border-b border-gray-200"
+                  >
+                    Actual
+                  </th>
+                  <th
+                    scope="col"
+                    className="p-4 text-right text-gray-700 font-semibold border-b border-gray-200"
+                  >
+                    Diff
+                  </th>
+                  <th
+                    scope="col"
+                    className="p-4 text-right text-gray-700 font-semibold border-b border-gray-200"
+                  >
+                    Audit State
+                  </th>
+                </tr>
+              </thead>
+              <tbody>
+                {Array.from({ length: 5 }).map((_, groupIndex) => (
+                  <React.Fragment key={groupIndex}>
+                    <tr className="border-b bg-gray-50">
+                      <td
+                        className="p-4 font-medium animate-pulse bg-gray-200 text-gray-200"
+                        colSpan={6}
+                      >
+                        <div className="h-4 w-48 bg-gray-200 rounded"></div>
+                      </td>
+                    </tr>
+                    {Array.from({ length: 3 }).map((__, rowIndex) => (
+                      <tr
+                        key={`${groupIndex}-${rowIndex}`}
+                        className="border-b hover:bg-gray-50 transition-colors"
+                      >
+                        <td className="p-4" />
+                        <td className="p-4">
+                          <div className="h-4 w-32 bg-gray-200 rounded animate-pulse mb-2" />
+                        </td>
+                        <td className="p-4 text-right">
+                          <div className="h-4 w-16 bg-gray-200 rounded animate-pulse ml-auto" />
+                        </td>
+                        <td className="p-4 text-right">
+                          <div className="h-4 w-16 bg-gray-200 rounded animate-pulse ml-auto" />
+                        </td>
+                        <td className="p-4 text-right">
+                          <div className="h-4 w-16 bg-gray-200 rounded animate-pulse ml-auto" />
+                        </td>
+                      </tr>
+                    ))}
+                  </React.Fragment>
+                ))}
+              </tbody>
+            </table>
+          </div>
+
+          {/* CARD SKELETON (Visible on mobile screens) */}
+          <div className="block sm:hidden space-y-4">
+            {Array.from({ length: 5 }).map((_, idx) => (
+              <div
+                key={idx}
+                className="bg-white border border-gray-200 rounded-lg p-4 animate-pulse"
+              >
+                <div className="h-4 w-32 bg-gray-200 rounded mb-4" />
+                {Array.from({ length: 3 }).map((__, rIdx) => (
+                  <div key={rIdx} className="flex justify-between mb-2">
+                    <div className="h-4 w-20 bg-gray-200 rounded" />
+                    <div className="h-4 w-16 bg-gray-200 rounded" />
+                  </div>
+                ))}
+              </div>
+            ))}
+          </div>
+        </>
+      ) : (
+        <>
+          {/* ---------- TABLE Layout (hidden on small screens) ---------- */}
+          <div className="hidden sm:block overflow-x-auto rounded-lg shadow-lg border border-gray-200">
+            <table
+              className="w-full border-collapse bg-white"
+              aria-label="Points Audit Table"
+            >
+              <thead>
+                <tr className="bg-gray-50">
+                  <th
+                    scope="col"
+                    className="p-4 text-left text-gray-700 font-semibold border-b border-gray-200 w-[60px]"
+                  ></th>
+                  <th
+                    scope="col"
+                    className="p-4 text-left text-gray-700 font-semibold border-b border-gray-200"
+                  >
+                    Strategy
+                  </th>
+                  <th
+                    scope="col"
+                    className="p-4 text-right text-gray-700 font-semibold border-b border-gray-200 w-[140px]"
+                  >
+                    Expected
+                  </th>
+                  <th
+                    scope="col"
+                    className="p-4 text-right text-gray-700 font-semibold border-b border-gray-200 w-[140px]"
+                  >
+                    Actual
+                  </th>
+                  <th
+                    scope="col"
+                    className="p-4 text-right text-gray-700 font-semibold border-b border-gray-200 w-[160px]"
+                  >
+                    Diff
+                  </th>
+                  <th
+                    scope="col"
+                    className="p-4 text-right text-gray-700 font-semibold border-b border-gray-200 w-[160px]"
+                  >
+                    <div>Audit State</div>
+                    <div className="text-xs font-normal">
+                      Last Snapshot, Diff%
+                    </div>
+                  </th>
+                </tr>
+              </thead>
+
+              <tbody>
+                {Object.entries(groupedData).map(([pointsId, rows]) => {
+                  // Friendly label fallback
+                  const displayName =
+                    pointIdToFriendlyName[pointsId] || pointsId;
+
+                  return (
+                    <React.Fragment key={pointsId}>
+                      {/* GROUP HEADER ROW */}
+                      <tr className="border-b bg-gray-100">
+                        <td className="p-4 font-semibold" colSpan={6}>
+                          {displayName}
+                        </td>
+                      </tr>
+
+                      {/* SUB ROWS */}
+                      {rows.map((row, subIndex) => {
+                        const actual = parseFloat(row.actualPoints);
+                        const expected = parseFloat(row.expectedPoints);
+                        const diff = actual - expected;
+                        const percentDiff =
+                          expected === 0 ? 0 : (diff / expected) * 100;
+
+                        const noExpectedPoints = row.expectedPoints === "0";
+
+                        const strategyConfig = config.find(
+                          (c) => c.strategy === row.strategy
+                        );
+
+                        const state = strategyConfig?.points.find(
+                          (p) => p.type === row.pointsId
+                        )?.state;
+
+                        return (
+                          <tr
+                            key={subIndex}
+                            className="border-b hover:bg-gray-50 transition-colors"
+                          >
+                            <td className="py-5 px-4" />
+                            <td className="py-5 px-4">
+                              <div className="font-medium text-gray-900">
+                                <span className="text-gray-900 truncate max-w-[180px]">
+                                  {row.strategy}
+                                </span>
+                                <InfoTooltip
+                                  externalAppURL={row.externalAppURL}
+                                  owner={row.owner}
+                                  dataSourceURLs={row.dataSourceURLs}
+                                  pointsBySource={row.pointsBySource}
+                                />
+                              </div>
+                            </td>
+                            {noExpectedPoints ? (
+                              <td className="py-5 px-2 bg-gray-50 text-right font-mono text-gray-600">
+                                N/A
+                              </td>
+                            ) : (
+                              <td className="py-5 px-2 text-right font-mono text-gray-600">
+                                {expected.toFixed(4)}
+                              </td>
+                            )}
+                            <td className="py-5 px-2 text-right font-mono text-gray-600">
+                              <Tooltip.Provider delayDuration={0}>
+                                <Tooltip.Root>
+                                  <Tooltip.Trigger asChild>
+                                    <span className="cursor-pointer border-b border-dotted border-gray-400 hover:border-indigo-500 transition-colors">
+                                      {actual.toLocaleString("en-US", {
+                                        minimumFractionDigits: 4,
+                                        maximumFractionDigits: 4,
+                                      })}
+                                    </span>
+                                  </Tooltip.Trigger>
+                                  <Tooltip.Portal>
+                                    <Tooltip.Content
+                                      className="bg-white border border-gray-200 rounded-lg shadow-lg p-4 z-50 min-w-[320px]"
+                                      sideOffset={5}
+                                    >
+                                      {historicalData[
+                                        `${row.strategy}-${row.pointsId}`
+                                      ] && (
+                                        <div>
+                                          <div className="flex items-center gap-2 mb-3">
+                                            <div className="h-2 w-2 rounded-full bg-indigo-500" />
+                                            <h3 className="text-sm font-semibold text-gray-900">
+                                              Historical Points
+                                            </h3>
+                                          </div>
+
+                                          <div className="space-y-2">
+                                            {historicalData[
+                                              `${row.strategy}-${row.pointsId}`
+                                            ]
+                                              .slice(0, 5)
+                                              .map((d, i) => (
+                                                <div
+                                                  key={i}
+                                                  className="flex justify-between items-center py-1.5 px-2 rounded hover:bg-gray-50"
+                                                >
+                                                  <div className="flex flex-col min-w-[140px]">
+                                                    <span className="text-xs font-medium text-gray-900">
+                                                      {new Date(
+                                                        d.created_at
+                                                      ).toLocaleDateString(
+                                                        "en-US",
+                                                        {
+                                                          dateStyle: "medium",
+                                                        }
+                                                      )}
+                                                    </span>
+                                                    <span className="text-xs text-gray-500">
+                                                      {new Date(
+                                                        d.created_at
+                                                      ).toLocaleTimeString(
+                                                        "en-US",
+                                                        {
+                                                          timeStyle: "short",
+                                                        }
+                                                      )}
+                                                    </span>
+                                                  </div>
+                                                  <span className="font-mono font-medium text-indigo-600 ml-4">
+                                                    {parseFloat(
+                                                      d.actualPoints
+                                                    ).toLocaleString("en-US", {
+                                                      minimumFractionDigits: 4,
+                                                      maximumFractionDigits: 4,
+                                                    })}
+                                                  </span>
+                                                </div>
+                                              ))}
+                                          </div>
+
+                                          {historicalData[
+                                            `${row.strategy}-${row.pointsId}`
+                                          ].length > 5 && (
+                                            <a
+                                              href="#"
+                                              onClick={(e) => {
+                                                e.preventDefault();
+                                                setSelectedHistory({
+                                                  isOpen: true,
+                                                  strategy: row.strategy,
+                                                  pointsId: row.pointsId,
+                                                });
+                                              }}
+                                              className="mt-3 text-xs text-indigo-600 hover:text-indigo-800 block text-center"
+                                            >
+                                              View{" "}
+                                              {historicalData[
+                                                `${row.strategy}-${row.pointsId}`
+                                              ].length - 5}{" "}
+                                              more entries
+                                            </a>
+                                          )}
+                                        </div>
+                                      )}
+                                      <Tooltip.Arrow className="fill-white" />
+                                    </Tooltip.Content>
+                                  </Tooltip.Portal>
+                                </Tooltip.Root>
+                              </Tooltip.Provider>
+                            </td>
+                            {noExpectedPoints ? (
+                              <td className="py-5 px-2 bg-gray-50"></td>
+                            ) : (
+                              <td
+                                className={`py-5 px-2 text-right font-mono font-medium whitespace-nowrap ${
+                                  Math.abs(percentDiff) < 0.1
+                                    ? "text-indigo-600 bg-indigo-50"
+                                    : diff < 0
+                                    ? "text-red-600 bg-red-50"
+                                    : "text-emerald-600 bg-emerald-50"
+                                }`}
+                              >
+                                {Math.abs(percentDiff) < 0.1 && "⭐"}{" "}
+                                {Math.abs(diff).toFixed(4)}
+                                <span className="text-xs ml-1">
+                                  ({percentDiff.toFixed(1)}%)
+                                </span>
+                              </td>
+                            )}
+                            {state && (
+                              <td className="py-5 px-2 text-right">
+                                <div className="flex items-center justify-end gap-1.5 font-medium font-mono">
+                                  <Image
+                                    src={`/${
+                                      {
+                                        verified: "check-circle",
+                                        delayed: "clock-rewind",
+                                        partial: "pie-chart",
+                                      }[state?.value?.toLowerCase() ?? ""]
+                                    }.svg`}
+                                    width="16"
+                                    height="16"
+                                    alt={state?.value ?? ""}
+                                  />
+                                  <span>{state?.value}</span>
+                                </div>
+                                <div className="text-gray-500 text-xs font-mono">
+                                  {state?.lastSnapshot}, {state?.diff}
+                                </div>
+                              </td>
+                            )}
+                          </tr>
+                        );
+                      })}
+                    </React.Fragment>
+                  );
+                })}
+              </tbody>
+            </table>
+          </div>
+
+          {/* ---------- CARD Layout (only shown on small screens) ---------- */}
+          <div className="block sm:hidden space-y-6">
             {Object.entries(groupedData).map(([pointsId, rows]) => {
-              // Friendly label fallback
+              // Friendly label
               const displayName = pointIdToFriendlyName[pointsId] || pointsId;
 
               return (
-                <React.Fragment key={pointsId}>
-                  {/* GROUP HEADER ROW */}
-                  <tr className="border-b bg-gray-100">
-                    <td className="p-4 font-semibold" colSpan={6}>
-                      {displayName}
-                    </td>
-                  </tr>
+                <div
+                  key={pointsId}
+                  className="bg-white border border-gray-200 rounded-lg p-4 shadow"
+                >
+                  {/* GROUP HEADER */}
+                  <h2 className="text-lg font-semibold mb-4">{displayName}</h2>
 
-                  {/* SUB ROWS */}
-                  {rows.map((row, subIndex) => {
-                    const actual = parseFloat(row.actualPoints);
-                    const expected = parseFloat(row.expectedPoints);
-                    const diff = actual - expected;
-                    const percentDiff =
-                      expected === 0 ? 0 : (diff / expected) * 100;
+                  {/* SUB ROWS AS CARDS */}
+                  <div className="space-y-2">
+                    {rows.map((row, subIndex) => {
+                      const actual = parseFloat(row.actualPoints);
+                      const expected = parseFloat(row.expectedPoints);
+                      const diff = actual - expected;
+                      const percentDiff =
+                        expected === 0 ? 0 : (diff / expected) * 100;
 
-                    const noExpectedPoints = row.expectedPoints === "0";
-
-                    const strategyConfig = CONFIG.find(
-                      (c) => c.strategy === row.strategy
-                    );
-
-                    const state = strategyConfig?.points.find(
-                      (p) => p.type === row.pointsId
-                    )?.state;
-
-                    return (
-                      <tr
-                        key={subIndex}
-                        className="border-b hover:bg-gray-50 transition-colors"
-                      >
-                        <td className="py-5 px-4" />
-                        <td className="py-5 px-4">
-                          <div className="font-medium text-gray-900">
+                      return (
+                        <div
+                          key={subIndex}
+                          className="p-3 rounded-md border border-gray-100"
+                        >
+                          <div className="flex items-center justify-between mb-1">
+                            <span className="font-medium text-gray-700">
+                              Strategy
+                            </span>
                             <span className="text-gray-900 truncate max-w-[180px]">
                               {row.strategy}
                             </span>
-                            <InfoTooltip
-                              externalAppURL={row.externalAppURL}
-                              owner={row.owner}
-                              dataSourceURLs={row.dataSourceURLs}
-                              pointsBySource={row.pointsBySource}
-                            />
                           </div>
-                        </td>
-                        {noExpectedPoints ? (
-                          <td className="py-5 px-2 bg-gray-50 text-right font-mono text-gray-600">
-                            N/A
-                          </td>
-                        ) : (
-                          <td className="py-5 px-2 text-right font-mono text-gray-600">
-                            {expected.toFixed(4)}
-                          </td>
-                        )}
-                        <td className="py-5 px-2 text-right font-mono text-gray-600">
-                          <Tooltip.Provider delayDuration={0}>
-                            <Tooltip.Root>
-                              <Tooltip.Trigger asChild>
-                                <span className="cursor-pointer border-b border-dotted border-gray-400 hover:border-indigo-500 transition-colors">
-                                  {actual.toLocaleString("en-US", {
-                                    minimumFractionDigits: 4,
-                                    maximumFractionDigits: 4,
-                                  })}
-                                </span>
-                              </Tooltip.Trigger>
-                              <Tooltip.Portal>
-                                <Tooltip.Content
-                                  className="bg-white border border-gray-200 rounded-lg shadow-lg p-4 z-50 min-w-[320px]"
-                                  sideOffset={5}
-                                >
-                                  {historicalData[
-                                    `${row.strategy}-${row.pointsId}`
-                                  ] && (
-                                    <div>
-                                      <div className="flex items-center gap-2 mb-3">
-                                        <div className="h-2 w-2 rounded-full bg-indigo-500" />
-                                        <h3 className="text-sm font-semibold text-gray-900">
-                                          Historical Points
-                                        </h3>
-                                      </div>
-
-                                      <div className="space-y-2">
-                                        {historicalData[
-                                          `${row.strategy}-${row.pointsId}`
-                                        ]
-                                          .slice(0, 5)
-                                          .map((d, i) => (
-                                            <div
-                                              key={i}
-                                              className="flex justify-between items-center py-1.5 px-2 rounded hover:bg-gray-50"
-                                            >
-                                              <div className="flex flex-col min-w-[140px]">
-                                                <span className="text-xs font-medium text-gray-900">
-                                                  {new Date(
-                                                    d.created_at
-                                                  ).toLocaleDateString(
-                                                    "en-US",
-                                                    {
-                                                      dateStyle: "medium",
-                                                    }
-                                                  )}
-                                                </span>
-                                                <span className="text-xs text-gray-500">
-                                                  {new Date(
-                                                    d.created_at
-                                                  ).toLocaleTimeString(
-                                                    "en-US",
-                                                    {
-                                                      timeStyle: "short",
-                                                    }
-                                                  )}
-                                                </span>
-                                              </div>
-                                              <span className="font-mono font-medium text-indigo-600 ml-4">
-                                                {parseFloat(
-                                                  d.actualPoints
-                                                ).toLocaleString("en-US", {
-                                                  minimumFractionDigits: 4,
-                                                  maximumFractionDigits: 4,
-                                                })}
-                                              </span>
-                                            </div>
-                                          ))}
-                                      </div>
-
-                                      {historicalData[
-                                        `${row.strategy}-${row.pointsId}`
-                                      ].length > 5 && (
-                                        <a
-                                          href="#"
-                                          onClick={(e) => {
-                                            e.preventDefault();
-                                            setSelectedHistory({
-                                              isOpen: true,
-                                              strategy: row.strategy,
-                                              pointsId: row.pointsId,
-                                            });
-                                          }}
-                                          className="mt-3 text-xs text-indigo-600 hover:text-indigo-800 block text-center"
-                                        >
-                                          View{" "}
-                                          {historicalData[
-                                            `${row.strategy}-${row.pointsId}`
-                                          ].length - 5}{" "}
-                                          more entries
-                                        </a>
-                                      )}
-                                    </div>
-                                  )}
-                                  <Tooltip.Arrow className="fill-white" />
-                                </Tooltip.Content>
-                              </Tooltip.Portal>
-                            </Tooltip.Root>
-                          </Tooltip.Provider>
-                        </td>
-                        {noExpectedPoints ? (
-                          <td className="py-5 px-2 bg-gray-50"></td>
-                        ) : (
-                          <td
-                            className={`py-5 px-2 text-right font-mono font-medium whitespace-nowrap ${
-                              Math.abs(percentDiff) < 0.1
-                                ? "text-indigo-600 bg-indigo-50"
-                                : diff < 0
-                                ? "text-red-600 bg-red-50"
-                                : "text-emerald-600 bg-emerald-50"
-                            }`}
-                          >
-                            {Math.abs(percentDiff) < 0.1 && "⭐"}{" "}
-                            {Math.abs(diff).toFixed(4)}
-                            <span className="text-xs ml-1">
-                              ({percentDiff.toFixed(1)}%)
+                          <div className="flex items-center justify-between mb-1">
+                            <span className="font-medium text-gray-700">
+                              Expected
                             </span>
-                          </td>
-                        )}
-                        {state && (
-                          <td className="py-5 px-2 text-right">
-                            <div className="flex items-center justify-end gap-1.5 font-medium font-mono">
-                              <Image
-                                src={`/${
-                                  {
-                                    verified: "check-circle",
-                                    delayed: "clock-rewind",
-                                    partial: "pie-chart",
-                                  }[state?.value?.toLowerCase() ?? ""]
-                                }.svg`}
-                                width="16"
-                                height="16"
-                                alt={state?.value ?? ""}
-                              />
-                              <span>{state?.value}</span>
-                            </div>
-                            <div className="text-gray-500 text-xs font-mono">
-                              {state?.lastSnapshot}, {state?.diff}
-                            </div>
-                          </td>
-                        )}
-                      </tr>
-                    );
-                  })}
-                </React.Fragment>
+                            <span className="font-mono text-gray-600">
+                              {expected.toFixed(4)}
+                            </span>
+                          </div>
+                          <div className="flex items-center justify-between mb-1">
+                            <span className="font-medium text-gray-700">
+                              Actual
+                            </span>
+                            <span className="font-mono text-gray-600">
+                              {actual.toFixed(4)}
+                            </span>
+                          </div>
+                          <div className="flex items-center justify-between">
+                            <span className="font-medium text-gray-700">
+                              Diff
+                            </span>
+                            <span
+                              className={`font-mono font-medium ${
+                                Math.abs(percentDiff) < 0.1
+                                  ? "text-indigo-600 bg-indigo-50 px-2 rounded"
+                                  : diff < 0
+                                  ? "text-red-600 bg-red-50 px-2 rounded"
+                                  : "text-emerald-600 bg-emerald-50 px-2 rounded"
+                              }`}
+                            >
+                              {Math.abs(diff).toFixed(4)} (
+                              {percentDiff.toFixed(1)}
+                              %)
+                              {Math.abs(percentDiff) < 0.1 && " ⭐"}
+                            </span>
+                          </div>
+                        </div>
+                      );
+                    })}
+                  </div>
+                </div>
               );
             })}
-          </tbody>
-        </table>
-      </div>
-
-      {/* ---------- CARD Layout (only shown on small screens) ---------- */}
-      <div className="block sm:hidden space-y-6">
-        {Object.entries(groupedData).map(([pointsId, rows]) => {
-          // Friendly label
-          const displayName = pointIdToFriendlyName[pointsId] || pointsId;
-
-          return (
-            <div
-              key={pointsId}
-              className="bg-white border border-gray-200 rounded-lg p-4 shadow"
-            >
-              {/* GROUP HEADER */}
-              <h2 className="text-lg font-semibold mb-4">{displayName}</h2>
-
-              {/* SUB ROWS AS CARDS */}
-              <div className="space-y-2">
-                {rows.map((row, subIndex) => {
-                  const actual = parseFloat(row.actualPoints);
-                  const expected = parseFloat(row.expectedPoints);
-                  const diff = actual - expected;
-                  const percentDiff =
-                    expected === 0 ? 0 : (diff / expected) * 100;
-
-                  return (
-                    <div
-                      key={subIndex}
-                      className="p-3 rounded-md border border-gray-100"
-                    >
-                      <div className="flex items-center justify-between mb-1">
-                        <span className="font-medium text-gray-700">
-                          Strategy
-                        </span>
-                        <span className="text-gray-900 truncate max-w-[180px]">
-                          {row.strategy}
-                        </span>
-                      </div>
-                      <div className="flex items-center justify-between mb-1">
-                        <span className="font-medium text-gray-700">
-                          Expected
-                        </span>
-                        <span className="font-mono text-gray-600">
-                          {expected.toFixed(4)}
-                        </span>
-                      </div>
-                      <div className="flex items-center justify-between mb-1">
-                        <span className="font-medium text-gray-700">
-                          Actual
-                        </span>
-                        <span className="font-mono text-gray-600">
-                          {actual.toFixed(4)}
-                        </span>
-                      </div>
-                      <div className="flex items-center justify-between">
-                        <span className="font-medium text-gray-700">Diff</span>
-                        <span
-                          className={`font-mono font-medium ${
-                            Math.abs(percentDiff) < 0.1
-                              ? "text-indigo-600 bg-indigo-50 px-2 rounded"
-                              : diff < 0
-                              ? "text-red-600 bg-red-50 px-2 rounded"
-                              : "text-emerald-600 bg-emerald-50 px-2 rounded"
-                          }`}
-                        >
-                          {Math.abs(diff).toFixed(4)} ({percentDiff.toFixed(1)}
-                          %)
-                          {Math.abs(percentDiff) < 0.1 && " ⭐"}
-                        </span>
-                      </div>
-                    </div>
-                  );
-                })}
-              </div>
-            </div>
-          );
-        })}
-      </div>
+          </div>
+        </>
+      )}
 
       {/* History Modal */}
       {selectedHistory.isOpen && (
